처방전 넣음

+++ app/src/main/java/com/techwork/kjc/mvp_project/util/Strandard2.java
@@ -77,7 +77,7 @@ public class Strandard2 {
     };
 
 
-    public static String evaluation(EvaluePart evaluePart, int val){
+    public static String evaluation(EvaluePart evaluePart, double val){
         EvalueClass evalueClass = standardMap.get(evaluePart);
         if(evalueClass == null) return None;
         if( val > evalueClass.top){
